Added multiple tests for subcourse() method of ORCPerformanceCurveCourse and implementation.

+++ java/com.sap.sailing.domain/src/com/sap/sailing/domain/orc/ORCPerformanceCurveCourse.java
@@ -15,5 +15,5 @@ public interface ORCPerformanceCurveCourse {
         return result;
     }
     
-    ORCPerformanceCurveCourse subcourse(int lastFinishedLeg, float perCentOfCurrentLeg);
+    ORCPerformanceCurveCourse subcourse(int lastFinishedLeg, double perCentOfCurrentLeg);
 }
